页面删除提示，以及保存拆分

+++ src/views/system/role/role.vue
@@ -145,16 +145,32 @@ export default {
       this.showDialogForm(row.roleId);
     },
     handleDelete (index, row) {
-      deleteRole(row.roleId).then(() => {
-        this.$message.success("删除成功");
-        this.searchList();
-      })
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteRole(row.roleId).then(() => {
+          this.$message.success("删除成功");
+          this.searchList();
+        })
+      }).catch(() => {
+      });
+
     },
     deleteByIds () {
-      deleteRole(this.roleIds).then(() => {
-        this.$message.success("删除成功");
-        this.searchList();
-      })
+      this.$confirm('此操作将永久删除, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteRole(this.roleIds).then(() => {
+          this.$message.success("删除成功");
+          this.searchList();
+        })
+      }).catch(() => {
+      });
+
     },
     handleSelectionChange (val) {
       this.roleIds = val.map(item => item.roleId);
